refactor(captp): less opinionated trapGuest (part 1/2)
